+++ configPlugins.js
@@ -1,0 +1,19 @@
+// untuk override status plugins
+
+// namaPlugins = true / false
+
+module.exports = {
+    status: {
+        bash: true,
+        debugJSON: true, parseMode: false,
+        foto: true, dokumen: true, video: true, audio: true, voice: true, sticker: true,
+        getMe: true, invoke: true,
+        pin: false, unpin: false,
+        ping: true, pong: true,
+        quotes: false, wikipedia: false,
+        uploadFoto: true, uploadDokumen: true, uploadVideo: true,
+        uploadAudio: true, uploadVoice: true, uploadSticker: true,
+        getUser: false,
+    },
+    
+}