+++ config_sample.js
@@ -1,10 +1,9 @@
+const plugin = require('./configPlugins')
+
 module.exports = {
   // di dapat dari https://my.telegram.org/
   API_ID: 123456,
   API_HASH: 'copasDiSini',
-
-  // path TDLib (tersedia untuk Linux 64 compile via Ubuntu 20.04)
-  pathTDLib: './tdlib/libtdjson.so',
 
   // aktifkan jika pakai bot API
   // jika bot API aktif, userbot otomatis OFF
@@ -16,6 +15,9 @@ module.exports = {
   // pake tipe string, format internasional 628xxxxxxx
 
   phone: '6281234567890', // ini contoh saja, ganti dengan yang sesuai
+
+  // path TDLib
+  pathTDLib: './tdlib/libtdjson.so',
 
   // untuk verbose mode
   debug: {
@@ -36,29 +38,18 @@ module.exports = {
 
   // detail aplikasi
 
+  // override plugins status, liat file pluginConfig.js
+  plugins: plugin.status,
+
   // untuk keperluan plugins
   base: {
     uploadPath: '',
-  },
-
-  // override plugins status
-  // namaPlugins = true / false
-  plugins: {
-    bash: false,
-    debugJSON: false, parseMode: false,
-    foto: true, dokumen: true, video: true, audio: true, voice: true, sticker: true,
-    getMe: true, invoke: false,
-    pin: true, unpin: true, ping: true, pong: true,
-    quotes: true, wikipedia: true,
-    uploadFoto: true, uploadDokumen: true, uploadVideo: true,
-    uploadAudio: true, uploadVoice: true, uploadSticker: true,
-    getUser: false
   },
 
   // userbot: jika true, maka pesan ditandai terbaca
   terbaca: true,
 
   // aktifkan jika pesan sendiri tidak ingin diproses
-  skipme: false,
-
+  skipMe: true,
+  skipMessage: true, // jika true, pas launching bot pesan lama ga akan direspon
 }